chore: add candidates to journey and add YAML files for republican GA governor race
