+++ app/content/candidates/ga/brad-raffensperger.yml
@@ -1,31 +1,31 @@
-fiveFifthsId: "candidate-brt-jones"
-name: "Burt Jones"
+fiveFifthsId: "brad-raffensperger"
+name: "Brad Raffensperger"
 state: "GA"
-race: "Governor"
+race: "Governor — Republican Primary 2026"
 party: "Republican"
 primary: "Republican Primary 2026"
 office_sought: "Governor"
 district: null
 region: "Statewide"
-incumbent: true
+incumbent: false
 debate_participant: true
-avatar_initials: "BJ"
-meta:
-  issues_addressed: 13
+avatar_initials: "BR"
+ballot_order: 6
+issues:
+  issues_addressed: 12
   issues_total: 13
-  clarity: "Specific — incumbent record cited"
+  clarity: "Specific on jobs and elections"
   sources_count: 3
   sources_list:
     - "Atlanta Press Club debate transcript (primary)"
-    - "burtjonesforga.com issues page"
+    - "bradforgeorgia.com issues pages"
     - "News coverage"
   last_updated: "2026-05-01"
-  callout: "Jones has the most detailed public record of any candidate, with an incumbent legislative history. Website confirms or adds positions on LGBTQ+ (Riley Gaines Act), fentanyl/criminal justice, rural healthcare, farm support, and workforce development. Most positions are confirmed by both debate statements and website."
   links:
     - label: "Campaign website"
-      url: "https://burtjonesforga.com/"
+      url: "https://bradforgeorgia.com/"
     - label: "Ballotpedia"
-      url: "https://ballotpedia.org/Burt_Jones"
+      url: "https://ballotpedia.org/Brad_Raffensperger"
 
 sections:
   - id: "economic-security"
@@ -38,53 +38,53 @@ sections:
         position_type: "pos"
         source: "both"
       - topic: "Taxation"
-        note: "Annual income tax rate cuts; moving toward 4.99% terminal rate by 2027"
+        note: "Cap property taxes; eliminate income tax; end non-essential property taxes for seniors"
         coverage: 3
         position_tag: "Reduce taxes"
         position_type: "pos"
         source: "both"
       - topic: "Small business & farming"
-        note: "Defend Georgia farmland; farm support; reduce permitting and licensing"
-        coverage: 3
-        position_tag: "Reduce regulation"
+        note: "Tax free zones for depressed communities; incentivize reshoring"
+        coverage: 2
+        position_tag: "Market-driven solutions"
         position_type: "pos"
         source: "both"
       - topic: "Workers & wages"
-        note: "Cut fees and burdensome regulations; vocational funding"
-        coverage: 2
-        position_tag: "Reduce labor regulation"
-        position_type: "pos"
-        source: "website"
+        note: "Minimum wage, labor protections"
+        coverage: 0
+        position_tag: null
+        position_type: "none"
+        source: null
 
   - id: "civil-rights-justice"
     title: "Civil rights & justice"
     items:
       - topic: "Criminal justice system"
-        note: "Mandatory minimums for gang offenders and fentanyl traffickers"
+        note: "Gangs, cartels, mandatory sentencing; end revolving-door justice"
         coverage: 3
         position_tag: "Increase enforcement"
         position_type: "pos"
         source: "both"
       - topic: "Reproductive policy"
-        note: "Voted for heartbeat bill 2019; supports exceptions for rape, incest, mother's health"
-        coverage: 3
-        position_tag: "Pro-life"
-        position_type: "pos"
-        source: "both"
+        note: "Abortion, contraception, medical decisions"
+        coverage: 0
+        position_tag: null
+        position_type: "none"
+        source: null
       - topic: "Voting & elections"
-        note: "SB 202, QR code removal, anti-sanctuary cities for elections"
+        note: "Defended GA elections record; first state citizenship verification"
         coverage: 3
         position_tag: "Strengthen voter verification"
         position_type: "pos"
         source: "both"
       - topic: "Gun policy"
-        note: "Ownership rights, regulations, safety"
+        note: "Explicitly states 'stand firm against any effort to undermine gun rights'"
         coverage: 2
         position_tag: "Expand gun rights"
         position_type: "pos"
         source: "website"
       - topic: "LGBTQ+ policy"
-        note: "Riley Gaines Act; bans biological males in female sports; separate facilities"
+        note: "Ban puberty blockers and gender surgeries for minors; ban biological males in female sports"
         coverage: 3
         position_tag: "Limit or oppose expansions"
         position_type: "pos"
@@ -94,23 +94,23 @@ sections:
     title: "Access & services"
     items:
       - topic: "Healthcare"
-        note: "Expanded rural access; easier healthcare facility permitting regardless of zip code"
-        coverage: 3
-        position_tag: "Mixed approach"
-        position_type: "mixed"
-        source: "both"
+        note: "Tax incentives for childcare; address healthcare costs broadly"
+        coverage: 2
+        position_tag: "Expand private options"
+        position_type: "pos"
+        source: "website"
       - topic: "Education"
-        note: "Georgia Promise Scholarship; ban CRT; vocational investment"
+        note: "School choice; double school safety grants to $100k; end woke curricula"
         coverage: 3
         position_tag: "Support school choice"
         position_type: "pos"
         source: "both"
       - topic: "Broadband & infrastructure"
         note: "Rural connectivity, roads, utilities"
-        coverage: 2
-        position_tag: "Public-private partnership"
-        position_type: "pos"
-        source: "website"
+        coverage: 0
+        position_tag: null
+        position_type: "none"
+        source: null
 
   - id: "community-environment"
     title: "Community & environment"
@@ -122,14 +122,14 @@ sections:
         position_type: "none"
         source: null
       - topic: "Immigration"
-        note: "Anti-sanctuary cities; deport criminal illegal immigrants; crack down on cartels"
+        note: "Partner with Trump admin; deport criminal aliens; fund state law enforcement"
         coverage: 3
         position_tag: "Stricter enforcement"
         position_type: "pos"
-        source: "both"
+        source: "website"
       - topic: "Housing"
-        note: "Affordability mentioned; rural farmland preservation"
+        note: "Repeal closing cost taxes; tax incentives to reduce cost burden"
         coverage: 2
-        position_tag: "Increase government support"
-        position_type: "pos"
+        position_tag: "Mixed approach"
+        position_type: "mixed"
         source: "website"
