+++ app/content/candidates/ga/burt-jones.yml
@@ -1,0 +1,136 @@
+fiveFifthsId: "burt-jones"
+name: "Burt Jones"
+state: "GA"
+race: "Governor — Republican Primary 2026"
+party: "Republican"
+primary: "Republican Primary 2026"
+office_sought: "Governor"
+district: null
+region: "Statewide"
+incumbent: false
+debate_participant: true
+avatar_initials: "BJ"
+ballot_order: 4
+issues:
+  issues_addressed: 13
+  issues_total: 13
+  clarity: "Specific — incumbent record cited"
+  sources_count: 3
+  sources_list:
+    - "Atlanta Press Club debate transcript (primary)"
+    - "burtjonesforga.com issues page"
+    - "News coverage"
+  last_updated: "2026-05-01"
+  callout: "Jones has the most detailed public record of any candidate, with an incumbent legislative history. Website confirms or adds positions on LGBTQ+ (Riley Gaines Act), fentanyl/criminal justice, rural healthcare, farm support, and workforce development. Most positions are confirmed by both debate statements and website."
+  links:
+    - label: "Campaign website"
+      url: "https://burtjonesforga.com/"
+    - label: "Ballotpedia"
+      url: "https://ballotpedia.org/Burt_Jones"
+
+sections:
+  - id: "economic-security"
+    title: "Economic security"
+    items:
+      - topic: "Cost of living"
+        note: "Prices, wages, groceries, energy"
+        coverage: 3
+        position_tag: "Market-driven solutions"
+        position_type: "pos"
+        source: "both"
+      - topic: "Taxation"
+        note: "Annual income tax rate cuts; moving toward 4.99% terminal rate by 2027"
+        coverage: 3
+        position_tag: "Reduce taxes"
+        position_type: "pos"
+        source: "both"
+      - topic: "Small business & farming"
+        note: "Defend Georgia farmland; farm support; reduce permitting and licensing"
+        coverage: 3
+        position_tag: "Reduce regulation"
+        position_type: "pos"
+        source: "both"
+      - topic: "Workers & wages"
+        note: "Cut fees and burdensome regulations; vocational funding"
+        coverage: 2
+        position_tag: "Reduce labor regulation"
+        position_type: "pos"
+        source: "website"
+
+  - id: "civil-rights-justice"
+    title: "Civil rights & justice"
+    items:
+      - topic: "Criminal justice system"
+        note: "Mandatory minimums for gang offenders and fentanyl traffickers"
+        coverage: 3
+        position_tag: "Increase enforcement"
+        position_type: "pos"
+        source: "both"
+      - topic: "Reproductive policy"
+        note: "Voted for heartbeat bill 2019; supports exceptions for rape, incest, mother's health"
+        coverage: 3
+        position_tag: "Pro-life"
+        position_type: "pos"
+        source: "both"
+      - topic: "Voting & elections"
+        note: "SB 202, QR code removal, anti-sanctuary cities for elections"
+        coverage: 3
+        position_tag: "Strengthen voter verification"
+        position_type: "pos"
+        source: "both"
+      - topic: "Gun policy"
+        note: "Ownership rights, regulations, safety"
+        coverage: 2
+        position_tag: "Expand gun rights"
+        position_type: "pos"
+        source: "website"
+      - topic: "LGBTQ+ policy"
+        note: "Riley Gaines Act; bans biological males in female sports; separate facilities"
+        coverage: 3
+        position_tag: "Limit or oppose expansions"
+        position_type: "pos"
+        source: "website"
+
+  - id: "access-services"
+    title: "Access & services"
+    items:
+      - topic: "Healthcare"
+        note: "Expanded rural access; easier healthcare facility permitting regardless of zip code"
+        coverage: 3
+        position_tag: "Mixed approach"
+        position_type: "mixed"
+        source: "both"
+      - topic: "Education"
+        note: "Georgia Promise Scholarship; ban CRT; vocational investment"
+        coverage: 3
+        position_tag: "Support school choice"
+        position_type: "pos"
+        source: "both"
+      - topic: "Broadband & infrastructure"
+        note: "Rural connectivity, roads, utilities"
+        coverage: 2
+        position_tag: "Public-private partnership"
+        position_type: "pos"
+        source: "website"
+
+  - id: "community-environment"
+    title: "Community & environment"
+    items:
+      - topic: "Environment & energy"
+        note: "Climate, clean water, energy production"
+        coverage: 0
+        position_tag: null
+        position_type: "none"
+        source: null
+      - topic: "Immigration"
+        note: "Anti-sanctuary cities; deport criminal illegal immigrants; crack down on cartels"
+        coverage: 3
+        position_tag: "Stricter enforcement"
+        position_type: "pos"
+        source: "both"
+      - topic: "Housing"
+        note: "Affordability mentioned; rural farmland preservation"
+        coverage: 2
+        position_tag: "Increase government support"
+        position_type: "pos"
+        source: "website"
