+++ app/content/candidates/ga/chris-carr.yml
@@ -1,7 +1,7 @@
-fiveFifthsId: "candidate-brd-raffensperger"
-name: "Brad Raffensperger"
+fiveFifthsId: "chris-carr"
+name: "Chris Carr"
 state: "GA"
-race: "Governor"
+race: "Governor — Republican Primary 2026"
 party: "Republican"
 primary: "Republican Primary 2026"
 office_sought: "Governor"
@@ -9,24 +9,23 @@ district: null
 region: "Statewide"
 incumbent: false
 debate_participant: true
-avatar_initials: "BR"
-meta:
-  issues_addressed: 12
+avatar_initials: "CC"
+ballot_order: 1
+issues:
+  issues_addressed: 10
   issues_total: 13
-  clarity: "Specific on jobs and elections"
+  clarity: "General — themes over specifics"
   sources_count: 3
   sources_list:
     - "Atlanta Press Club debate transcript (primary)"
-    - "bradforgeorgia.com issues pages"
+    - "carrforgeorgia.com issues pages"
     - "News coverage"
   last_updated: "2026-05-01"
-  callout: "Website is one of the most policy-detailed in the field. Raffensperger explicitly states positions on LGBTQ+ (ban puberty blockers, ban biological males in female sports), immigration (support Trump deportation), eliminate income tax, expand school choice, and oppose gambling. Strong confirmation across both sources on elections and jobs."
-  data_note: "JSON source file contained a Ballotpedia link swap between Raffensperger and Williams — flagged for correction."
   links:
     - label: "Campaign website"
-      url: "https://bradforgeorgia.com/"
+      url: "https://carrforgeorgia.com/"
     - label: "Ballotpedia"
-      url: "https://ballotpedia.org/Brad_Raffensperger"
+      url: "https://ballotpedia.org/Chris_Carr_(Georgia)"
 
 sections:
   - id: "economic-security"
@@ -34,22 +33,22 @@ sections:
     items:
       - topic: "Cost of living"
         note: "Prices, wages, groceries, energy"
-        coverage: 3
+        coverage: 2
         position_tag: "Market-driven solutions"
         position_type: "pos"
         source: "both"
       - topic: "Taxation"
-        note: "Cap property taxes; eliminate income tax; end non-essential property taxes for seniors"
+        note: "Tax rates, cuts, who pays"
         coverage: 3
         position_tag: "Reduce taxes"
         position_type: "pos"
         source: "both"
       - topic: "Small business & farming"
-        note: "Tax free zones for depressed communities; incentivize reshoring"
+        note: "Support, capital access, regulation"
         coverage: 2
-        position_tag: "Market-driven solutions"
+        position_tag: "Reduce regulation"
         position_type: "pos"
-        source: "both"
+        source: "website"
       - topic: "Workers & wages"
         note: "Minimum wage, labor protections"
         coverage: 0
@@ -61,7 +60,7 @@ sections:
     title: "Civil rights & justice"
     items:
       - topic: "Criminal justice system"
-        note: "Gangs, cartels, mandatory sentencing; end revolving-door justice"
+        note: "Policing, sentencing, incarceration"
         coverage: 3
         position_tag: "Increase enforcement"
         position_type: "pos"
@@ -73,45 +72,45 @@ sections:
         position_type: "none"
         source: null
       - topic: "Voting & elections"
-        note: "Defended GA elections record; first state citizenship verification"
-        coverage: 3
+        note: "ID laws, ballot access, integrity"
+        coverage: 2
         position_tag: "Strengthen voter verification"
         position_type: "pos"
-        source: "both"
+        source: "debate"
       - topic: "Gun policy"
-        note: "Explicitly states 'stand firm against any effort to undermine gun rights'"
+        note: "Ownership rights, regulations, safety"
         coverage: 2
         position_tag: "Expand gun rights"
         position_type: "pos"
         source: "website"
       - topic: "LGBTQ+ policy"
-        note: "Ban puberty blockers and gender surgeries for minors; ban biological males in female sports"
-        coverage: 3
-        position_tag: "Limit or oppose expansions"
-        position_type: "pos"
-        source: "website"
+        note: "Protections, discrimination, parental rights"
+        coverage: 0
+        position_tag: null
+        position_type: "none"
+        source: null
 
   - id: "access-services"
     title: "Access & services"
     items:
       - topic: "Healthcare"
-        note: "Tax incentives for childcare; address healthcare costs broadly"
+        note: "Coverage, Medicaid, rural access"
         coverage: 2
-        position_tag: "Expand private options"
-        position_type: "pos"
+        position_tag: "Mixed approach"
+        position_type: "mixed"
         source: "website"
       - topic: "Education"
-        note: "School choice; double school safety grants to $100k; end woke curricula"
+        note: "Public schools, choice, funding, HBCUs"
         coverage: 3
-        position_tag: "Support school choice"
+        position_tag: "Both public and choice"
         position_type: "pos"
         source: "both"
       - topic: "Broadband & infrastructure"
         note: "Rural connectivity, roads, utilities"
-        coverage: 0
-        position_tag: null
-        position_type: "none"
-        source: null
+        coverage: 2
+        position_tag: "Increase public investment"
+        position_type: "pos"
+        source: "debate"
 
   - id: "community-environment"
     title: "Community & environment"
@@ -123,14 +122,14 @@ sections:
         position_type: "none"
         source: null
       - topic: "Immigration"
-        note: "Partner with Trump admin; deport criminal aliens; fund state law enforcement"
-        coverage: 3
+        note: "Enforcement, pathways, local policy"
+        coverage: 2
         position_tag: "Stricter enforcement"
         position_type: "pos"
         source: "website"
       - topic: "Housing"
-        note: "Repeal closing cost taxes; tax incentives to reduce cost burden"
+        note: "Affordability, development, tenant policy"
         coverage: 2
         position_tag: "Mixed approach"
         position_type: "mixed"
-        source: "website"
+        source: "both"
