+++ app/content/elections/ga/2026-midterms-primary.yml
@@ -40,4 +40,3 @@ voting:
 
 races:
   - "Governor — Republican Primary 2026"
-  - "Governor — Democratic Primary 2026"
