+++ app/pages/journey/election/candidates/[id].vue
@@ -3,19 +3,24 @@ import { AlertTriangle } from "lucide-vue-next";
 
 // These could come from route params or props depending on your routing setup
 const state = "GA";
-const race = "Governor — Republican Primary 2026";
+const route = useRoute();
+const race = route.params.id as string;
 
 const { data: candidates, status } = await useAsyncData(
-  `candidates-${state}`,
+  `candidates-${race}`,
   () =>
     queryCollection("candidates")
-      .where("state", "=", state)
+      .where("race", "=", race)
+      .order("ballot_order", "ASC")
       .all(),
 );
 </script>
 
 <template>
-  <div class="max-w-3xl mx-auto px-4 py-6">
+  <div class="max-w-3xl mx-auto px-4 py-6 mt-8">
+    <PageTitle>
+      Candidates
+    </PageTitle>
     <div v-if="status === 'pending'" class="flex justify-center py-16">
       <span class="loading loading-spinner loading-lg" />
     </div>
