+++ app/pages/journey/election/candidates/score/[id].vue
