--- content/candidates/ga/tom-williams.yml
@@ -1,133 +1,0 @@
-fiveFifthsId: "candidate-tom-williams"
-name: "Tom Williams"
-state: "GA"
-race: "Governor"
-party: "Republican"
-primary: "Republican Primary 2026"
-office_sought: "Governor"
-district: null
-region: "Statewide"
-incumbent: false
-debate_participant: true
-avatar_initials: "TW"
-meta:
-  issues_addressed: 9
-  issues_total: 13
-  clarity: "Specific on elections only in debate"
-  sources_count: 2
-  sources_list:
-    - "Atlanta Press Club debate transcript (primary)"
-    - "tomwilliamsforga.com issues page"
-  last_updated: "2026-05-01"
-  callout: "Website fills major gaps the debate left open. Williams explicitly states: pro-life position, school choice, Second Amendment defense, opposition to transgender policy in schools and sports, and lower income taxes. His debate focus was almost entirely elections and campaign finance; the website shows a broader platform."
-  data_note: "Williams asked explicitly to be called 'Tom Williams' not 'Thomas Williams' during the debate."
-  links:
-    - label: "Campaign website"
-      url: "https://tomwilliamsforga.com/"
-
-sections:
-  - id: "economic-security"
-    title: "Economic security"
-    items:
-      - topic: "Cost of living"
-        note: "Attributes affordability crisis to radical Democrats"
-        coverage: 1
-        position_tag: "Market-driven solutions"
-        position_type: "pos"
-        source: "both"
-      - topic: "Taxation"
-        note: "Fight to reduce state income taxes"
-        coverage: 2
-        position_tag: "Reduce taxes"
-        position_type: "pos"
-        source: "website"
-      - topic: "Small business & farming"
-        note: "Support, capital access, regulation"
-        coverage: 0
-        position_tag: null
-        position_type: "none"
-        source: null
-      - topic: "Workers & wages"
-        note: "Minimum wage, labor protections"
-        coverage: 0
-        position_tag: null
-        position_type: "none"
-        source: null
-
-  - id: "civil-rights-justice"
-    title: "Civil rights & justice"
-    items:
-      - topic: "Criminal justice system"
-        note: "Defend law enforcement from defunding; oppose rogue prosecutors"
-        coverage: 2
-        position_tag: "Increase enforcement"
-        position_type: "pos"
-        source: "website"
-      - topic: "Reproductive policy"
-        note: "'Protect every precious unborn child'; strongest pro-life laws in country"
-        coverage: 3
-        position_tag: "Pro-life"
-        position_type: "pos"
-        source: "website"
-      - topic: "Voting & elections"
-        note: "Paper ballots, no QR codes, outlaw leadership PACs, ban money in politics"
-        coverage: 3
-        position_tag: "Strengthen voter verification"
-        position_type: "pos"
-        source: "both"
-      - topic: "Gun policy"
-        note: "Defend Second Amendment against local authorities who try to weaken it"
-        coverage: 2
-        position_tag: "Expand gun rights"
-        position_type: "pos"
-        source: "website"
-      - topic: "LGBTQ+ policy"
-        note: "Protect women's sports, locker rooms, restrooms from biological males"
-        coverage: 2
-        position_tag: "Limit or oppose expansions"
-        position_type: "pos"
-        source: "website"
-
-  - id: "access-services"
-    title: "Access & services"
-    items:
-      - topic: "Healthcare"
-        note: "Coverage, Medicaid, rural access"
-        coverage: 0
-        position_tag: null
-        position_type: "none"
-        source: null
-      - topic: "Education"
-        note: "School choice; protect parental rights; sanction schools that violate them"
-        coverage: 3
-        position_tag: "Support school choice"
-        position_type: "pos"
-        source: "website"
-      - topic: "Broadband & infrastructure"
-        note: "Rural connectivity, roads, utilities"
-        coverage: 0
-        position_tag: null
-        position_type: "none"
-        source: null
-
-  - id: "community-environment"
-    title: "Community & environment"
-    items:
-      - topic: "Environment & energy"
-        note: "Climate, clean water, energy production"
-        coverage: 0
-        position_tag: null
-        position_type: "none"
-        source: null
-      - topic: "Immigration"
-        note: "Strong borders; enforcement against criminal illegal aliens"
-        coverage: 2
-        position_tag: "Stricter enforcement"
-        position_type: "pos"
-        source: "website"
-      - topic: "Housing"
-        note: "Affordability, development, tenant policy"
-        coverage: 0
-        position_tag: null
-        position_type: "none"
-        source: null
